+++ web/documentserver-example/php-laravel/app/OnlyOffice/Managers/SettingsManager.php
@@ -16,12 +16,19 @@
  * limitations under the License.
  */
 
-namespace App\Services;
+namespace App\OnlyOffice\Managers;
 
-class ServerConfig extends Config
+use Exception;
+use Onlyoffice\DocsIntegrationSdk\Manager\Settings\SettingsManager as OnlyOfficeSettingsManager;
+
+class SettingsManager extends OnlyOfficeSettingsManager
 {
+    private array $config;
+
     public function __construct()
     {
+        parent::__construct();
+
         $publicServerUrl = rtrim(env('DOCUMENT_SERVER_PUBLIC_URL', 'http://documentserver'), '/');
         $privateServerUrl = rtrim(env('DOCUMENT_SERVER_PRIVATE_URL', $publicServerUrl), '/');
         $apiUrl = $publicServerUrl.'/'.env('DOCUMENT_SERVER_API_PATH', 'web-apps/apps/api/documents/api.js');
@@ -30,8 +37,14 @@ class ServerConfig extends Config
         $commandUrl = $privateServerUrl.'/'.env('DOCUMENT_SERVER_COMMAND_PATH', 'command');
         $jwtSecret = env('DOCUMENT_SERVER_JWT_SECRET', 'secret');
         $jwtUseForRequest = env('DOCUMENT_SERVER_JWT_USE_FOR_REQUEST', true);
+        $publicStorageUrl = rtrim(env('DOCUMENT_STORAGE_PUBLIC_URL', request()->schemeAndHttpHost()), '/');
+        $privateStorageUrl = rtrim(env('DOCUMENT_STORAGE_PRIVATE_URL', $publicStorageUrl), '/');
 
         $this->config = [
+            'documentServerInternalUrl' => $privateServerUrl,
+            'jwtKey' => $jwtSecret,
+            'jwtHeader' => env('DOCUMENT_SERVER_JWT_HEADER', 'Authorization'),
+            'jwtPrefix' => env('DOCUMENT_SERVER_JWT_HEADER', 'Bearer '),
             'conversion' => [
                 'timeout' => env('DOCUMENT_SERVER_CONVERSION_TIMEOUT', 120 * 1000),
                 'url' => $conversionUrl,
@@ -47,11 +60,48 @@ class ServerConfig extends Config
             ],
             'url' => [
                 'api' => $apiUrl,
-                'public' => $publicServerUrl,
-                'private' => $privateServerUrl,
                 'preloader' => $preloaderUrl,
                 'command' => $commandUrl,
+                'server' => [
+                    'public' => $publicServerUrl,
+                    'private' => $privateServerUrl,
+                ],
+                'storage' => [
+                    'private' => $privateStorageUrl,
+                    'public' => $publicStorageUrl,
+                ],
+            ],
+            'file' => [
+                'max_size' => env('DOCUMENT_STORAGE_MAXIMUM_FILE_SIZE', 5 * 1024 * 1024),
             ],
         ];
+    }
+
+    public function getServerUrl()
+    {
+        return $this->get('url.server.public');
+    }
+
+    public function getSetting($settingName)
+    {
+        return $this->get($settingName);
+    }
+
+    public function setSetting($settingName, $value, $createSetting = false) {}
+
+    private function get(string $key, mixed $default = null): mixed
+    {
+        $keys = explode('.', $key);
+        $result = $this->config;
+
+        try {
+            foreach ($keys as $key) {
+                $result = $result[$key];
+            }
+        } catch (Exception $e) {
+            $result = $default;
+        }
+
+        return $result;
     }
 }