+++ web/documentserver-example/php-laravel/app/UseCases/Docs/Command/UpdateMetaCommand.php
@@ -2,21 +2,15 @@
 
 namespace App\UseCases\Docs\Command;
 
-use App\Services\Docs\Command\CommandRequest;
+use App\OnlyOffice\Miscellaneous\CommandRequest;
 
 class UpdateMetaCommand
 {
     public function __invoke(UpdateMetaRequest $request): void
     {
-        $content = [
-            'c' => 'meta',
-            'key' => $request->key,
-            'meta' => [
-                'title' => $request->title,
-            ],
-        ];
+        $meta = ['title' => $request->title];
 
         app(CommandRequest::class)
-            ->send($content, $request->key);
+            ->updateMeta($request->key, $meta);
     }
 }