+++ web/documentserver-example/php-laravel/app/UseCases/Document/Find/FindAllDocumentsQueryHandler.php
@@ -3,15 +3,15 @@
 namespace App\UseCases\Document\Find;
 
 use App\Helpers\Path\PathInfo;
+use App\OnlyOffice\Managers\FormatManager;
 use App\Repositories\FileRepository;
-use App\Repositories\FormatRepository;
 use App\Repositories\VersionRepository;
 
 class FindAllDocumentsQueryHandler
 {
     public function __construct(
         private FileRepository $fileRepository,
-        private FormatRepository $formatRepository,
+        private FormatManager $formatManager,
         private VersionRepository $versionRepository,
     ) {}
 
@@ -27,7 +27,7 @@ class FindAllDocumentsQueryHandler
             $result[] = [
                 'filename' => PathInfo::basename($file->filename),
                 'version' => $currentVersion,
-                'format' => $this->formatRepository->find(PathInfo::extension($file->filename)),
+                'format' => $this->formatManager->find(PathInfo::extension($file->filename)),
                 'lastModified' => $file->modified,
             ];
         }