+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - php-laravel: create, edit, and submit pdf forms
 - php-laravel: show forgotten files on a seperate page
 - php-laravel: fetch files
+- php-laravel: integrate sdk
 - restore by url
 - refresh config
 - on uploading xml convert to supported type only